Move charge deposition and interpolation
