+++ wake_t/physics_models/plasma_wakefields/qs_rz_baxevanis/wakefield.py
@@ -3,7 +3,7 @@ import scipy.constants as ct
 import aptools.plasma_accel.general_equations as ge
 
 from .solver import calculate_wakefields
-from wake_t.interpolation import (
+from wake_t.particles.interpolation import (
     gather_field_cyl_linear, gather_main_fields_cyl_linear)
 from wake_t.utilities.other import generate_field_diag_dictionary
 from wake_t.physics_models.plasma_wakefields.base_wakefield import Wakefield
